Modularize code in main.js
Broke out logical segments of code into separate files, stored
in the components folder, for sanity.

Each component has a bind function that is used to connect to
the main ipc module when the app is loaded.
