+++ main.js
@@ -1,694 +1,35 @@
 if (require('electron-squirrel-startup')) return;
 
+/** Setup Globals */
 const electron = require('electron')
 const app = electron.app // Module to control application life.
-process.env.directory = process.env.directory || app.getAppPath()
 const BrowserWindow = electron.BrowserWindow // Module to create native browser window.
-const ipc = electron.ipcMain
-const dialog = electron.dialog
 const Menu = electron.Menu
 
+/** Set the current directory */
+process.env.directory = process.env.directory || app.getAppPath()
+
+/** Configure Monitoring Station server and initialize settings */
 require('./server')
-var settings = require('./helpers/settings')
-var workspaces = require('./helpers/workspaces')
-var ServerEvents = require('./helpers/server-events')
-ServerEvents.on('cl_upload_progress', function(done, bytes) {
-  if (done)
-    mainWindow.send('cl_upload_progress', { status: 'waiting' });
-  else {
-    var value = bytes, units = 'B';
-    if (bytes < 1024) {
-      //Do nothing
-    } else if (bytes < (1024 * 1024)) {
-      value = bytes / 1024;
-      units = 'KB';
-    } else {
-      value = bytes / (1024 * 1024);
-      units = 'MB';
-    }
-    value = +value.toFixed(2);
-    mainWindow.send('cl_upload_progress', { status: 'uploading', progress: value + ' ' + units });
-  }
-});
+const settings = require('./helpers/settings')
+const i18n = require('./helpers/locale.js')
 
-var http = require('http')
-var fs = require('fs-extra')
-var glob = require('glob')
-var readline = require('readline')
-var path = require('path')
-var unzip = require('unzip2');
-var watcher = require('chokidar');
-var GeoJson = require('./helpers/rebuild-geojson')
-var async = require('async');
-var i18n = require('./helpers/locale.js');
-var Tiles = require('./helpers/tile-layers.js');
-
-var TilesWatcherTimeout = null;
-var TilesWatcher = null;
-
-ipc.on('show_configuration', function (event, arg) {
-  try {
-    var _defaults = settings.getDefaults()
-    var _results = {
-      'directory': settings.getDataDirectory() || _defaults.data_directory,
-      'station': settings.getStation() || _defaults.station,
-      'baseUrl': settings.getBaseUrl(),
-      'shared_secret': settings.getSharedSecret(),
-      'locale': settings.getLocale() || _defaults.locale,
-      'community_lands': !(settings.getCommunityLandsServer() === undefined || settings.getCommunityLandsToken() === undefined),
-      'tiles': settings.getDefaultTilesDirectory(),
-      'tracks': settings.getTracksDirectory()
-    }
-    console.log(_results)
-    event.sender.send('has_configuration', _results)
-  } catch (err) {
-    event.sender.send('has_configuration', [])
-  }
-})
-
-ipc.on('backup_submissions', function (event, arg) {
-  var options = {
-    hostname: 'localhost',
-    port: settings.getPort() || 3000,
-    path: '/save/all',
-    method: 'GET'
-  }
-  http.request(options, function (res) {
-    var data = ''
-    res.on('data', function (chunk) {
-      data += chunk
-    })
-    res.on('end', function () {
-      result = JSON.parse(data);
-      result['cb'] = arg;
-      result['cancelled'] = false;
-      if (arg) {
-        options = {
-          defaultPath: result.location
-        }
-        dialog.showSaveDialog(mainWindow, options, function(filename) {
-          if (filename && result.location != filename) {
-            fs.copy(result.location, filename, function(err) {
-              if (err) {
-                result['error'] = true;
-                result['code'] = 'backup_failed';
-              } else {
-                result.location = filename;
-              }
-              event.sender.send('backup_submissions_complete', result);
-            });
-          } else {
-            result['cancelled'] = result.location == filename;
-            event.sender.send('backup_submissions_complete', result);
-          }
-        });
-      } else {
-        event.sender.send('backup_submissions_complete', result)
-      }
-    })
-  }).on('error', function (e) {
-    event.sender.send('backup_submissions_complete',{"error":true, "code":"could_not_connect", "message":"Could not connect to server"})
-  }).end()
-})
-
-ipc.on('check_last_backup', function (event, arg) {
-  var options = {
-    hostname: 'localhost',
-    port: settings.getPort() || 3000,
-    path: '/save/status',
-    method: 'GET'
-  }
-  http.request(options, function (res) {
-    var data = ''
-    res.on('data', function (chunk) {
-      data += chunk
-    })
-    res.on('end', function () {
-      event.sender.send('has_last_backup', data)
-    })
-  }).end()
-})
-
-ipc.on('form_delete', function (event, arg) {
-  var folder = settings.getUserFormsDirectory()
-  fs.readdir(folder, function (err, files) {
-    if (err) {
-      event.sender.send('has_form_delete')
-    } else {
-      var found = false
-      for (var i in files) {
-        if (files[i] === arg) {
-          fs.unlinkSync(path.join(folder, files[i]))
-          event.sender.send('has_form_delete')
-          found = true
-          break
-        }
-      }
-      if (!found)
-        event.sender.send('has_form_delete')
-    }
-  })
-})
-
-/*
- * Cheating a bit here -- instead of reading the entire file and parsing the
- * XML, going to read line-by-line instead until I find the interesting line,
- * then break early if possible.
- */
-function createFormReader(key) {
-  return function(cb) {
-    var reader = readline.createInterface({
-      input: fs.createReadStream(path.join(settings.getUserFormsDirectory(), key)),
-      terminal: true
-    });
-    var item = { file: key, name: key };
-    reader.on('line', function (input) {
-      var line = input.trim();
-      if (line.startsWith("<h:title")) {
-        item.name = line.substring("<h:title>".length, line.indexOf("</h:title>"));
-        /*
-         * FIXME: Hack to close the input stream early and stop reading extra
-         * information. Would like a better solution. For now, simulate terminal
-         * input of Ctrl+C/D
-         */
-        reader.write(null, { ctrl: true, name: require('os').platform() == 'win32' ? 'd' : 'c' })
-      }
-    });
-    reader.on('close', function () {
-      cb(null, item);
-    });
-  }
+/** Add Components to support electron app */
+const context = {
+  ipc: electron.ipcMain,
+  settings: settings,
+  getMainWindow: function() { return mainWindow; }
 }
+require('./components/filters').bind(context);
+require('./components/forms').bind(context);
+require('./components/import_export').bind(context);
+require('./components/settings').bind(context);
+require('./components/tracks').bind(context);
+require('./components/upload').bind(context);
+const Tiles = require('./components/tiles'); //We need this later
+Tiles.bind(context);
 
-ipc.on('form_list', function (event, arg) {
-  var folder = settings.getUserFormsDirectory();
-  fs.readdir(folder, function (err, files) {
-    var data = { forms: [] }
-    if (err) {
-      event.sender.send('has_form_list', data)
-    } else {
-      var parallels = files.map(createFormReader);
-      async.parallel(parallels, function(a_err, results) {
-        var data = { forms: [] }
-        if (!a_err) {
-          for (var key in results)
-            data.forms.push(results[key])
-          data.forms.sort(function(a, b) {
-            var l = a.name.toLowerCase();
-            var r = b.name.toLowerCase();
-            return l < r ? -1 : l > r ? 1 : 0;
-          });
-        }
-        event.sender.send('has_form_list', data);
-      });
-    }
-  })
-})
-
-function listTiles(cb) {
-  Tiles.get(function(err, md) {
-    if (err) {
-      /*
-       * Probably in cases of major error, should reset any custom installed 
-       * tile directories and start fresh
-       */
-      cb(err)
-    } else {
-      var warnFiles = warnFolders = false;
-      var result = { error: false, tiles: [], warnings: [], directories: [] };
-      Object.keys(md).forEach(function(file) {
-        var stats = md[file];
-        if (!stats.eligible)
-          warnFiles = true;
-        else {
-          result.tiles.push(stats);
-        }
-      });
-      if (warnFiles)
-        result.warnings.push("error.files_in_tiles_folder")
-      result.tiles.sort(function(a, b) { return a.name < b.name ? -1 : a.name > b.name ? 1 : 0 });
-      result.directories = settings.getTilesDirectories().map(function(value) {
-        var available = false;
-        try {
-          fs.accessSync(value, fs.F_OK | fs.R_OK); //only need to read
-          available = true;
-        } catch (e) { //Directory unavailable
-        }
-        return {
-          available: available,
-          directory: value,
-          isDefault: value == settings.getDefaultTilesDirectory()
-        }
-      });
-
-      cb(null, result);
-    }
-  });
-}
-
-ipc.on('tiles_list', function (event, arg) {
-  listTiles(function(err, result) {
-    if (!err)
-      event.sender.send('has_tiles_list', result);
-  });
-});
-
-ipc.on('select_data_directory', function (event, arg) {
-  var options = {
-    properties: ['openDirectory'],
-    defaultPath: settings.getDataDirectory()
-  }
-  dialog.showOpenDialog(mainWindow, options, function (folder) {
-    if (folder) {
-      event.sender.send('has_select_data_directory', folder)
-    }
-  })
-})
-
-ipc.on('select_form', function (event, arg) {
-  var options = {
-    properties: ['openFile', 'multiSelections'],
-    filters: [ { name: 'XML', extensions: ['xml'] } ]
-  }
-  dialog.showOpenDialog(mainWindow, options, function (arr) {
-    if (arr !== undefined) {
-      var destDir = settings.getUserFormsDirectory()
-      var uploaded = {
-        count: arr.length,
-        names: []
-      }
-      for (var index in arr) {
-        var source = arr[index]
-        var filename = path.basename(source)
-        var target = path.join(destDir, filename)
-        fs.copySync(source, target)
-        uploaded.names.push(filename)
-      }
-      event.sender.send('has_select_form', uploaded)
-    }
-  })
-})
-
-ipc.on('select_track_data', function (event, arg) {
-  var options = {
-    properties: ['openFile', 'multiSelections'],
-    filters: [ { name: 'ZIP', extensions: ['zip'] } ]
-  }
-  dialog.showOpenDialog(mainWindow, options, function (arr) {
-    if (arr !== undefined) {
-      var destDir = settings.getTracksDirectory()
-      var parallels = arr.map(function (value) {
-        return unzipTrackData(value, destDir);
-      });
-      async.parallel(parallels, function (err, result) {
-        event.sender.send('has_select_track_data', { errors: false })
-      });
-    }
-  })
-})
-
-function unzipTrackData(source, target) {
-  return function (cb) {
-    fs.createReadStream(source)
-      .pipe(unzip.Extract({ path: target }))
-      .on('close', function() {
-        cb()
-      });
-  }
-}
-
-ipc.on('tiles_add_directory', function (event) {
-  var options = {
-    properties: ['openDirectory'],
-    defaultPath: settings.getDataDirectory()
-  }
-  dialog.showOpenDialog(mainWindow, options, function (folders) {
-    if (folders) {
-      var folder = folders[0];
-      var existing = settings.getTilesDirectories();
-      if (existing.indexOf(folder) >= 0)
-        event.sender.send('show_status', { message: 'errors.duplicate_tiles_folder', options: { type: 'error' } });
-      else {
-        var current = process.env.tiles_directories;
-        if (!current)
-          current = folder;
-        else
-          current += ';' + folder;
-        settings.patch({ tiles_directories: current }, function (err) {
-          if (!err) {
-            process.env.tiles_directories = current;
-            setupTilesWatcher();
-            refreshTiles();
-          }
-        });
-      }
-    }
-  });
-});
-
-ipc.on('tiles_remove_directory', function (event, args) {
-  var current = (process.env.tiles_directories || '').split(";")
-  var idx = current.indexOf(args.directory)
-  if (idx > -1)
-    delete current[idx]
-  var result = ''
-  for (var i = 0; i < current.length; i++) {
-    if (current[i]) {
-      if (!result)
-        result = current[i]
-      else
-        result += ';' + current[i]
-    }
-  }
-  settings.patch({ tiles_directories: result }, function (err) {
-    if (!err) {
-      process.env.tiles_directories = result;
-      setupTilesWatcher();
-      refreshTiles();
-    }
-  });
-});
-
-ipc.on('tracks_list', function (event, arg) {
-  var pattern = path.join(settings.getTracksDirectory(), "**", "*.+(gpx|csv)");
-  glob(pattern, { matchBase: true, nodir: true, nocase: true }, function(err, matches) {
-    var result = {}
-    if (!err) {
-      for (var i = 0; i < matches.length; i++) {
-        var dir = matches[i];
-        var file = path.basename(dir);
-        var folder = path.basename(path.dirname(dir));
-        var list = result[folder] || []
-        list.push(file)
-        result[folder] = list
-      }
-    }
-    event.sender.send('has_tracks_list', result);
-  });
-});
-
-ipc.on('filter_list', function (event, arg) {
-  var options = {
-    hostname: 'localhost',
-    port: settings.getPort() || 3000,
-    path: '/mapfilter/filters',
-    method: 'GET'
-  }
-  http.request(options, function (res) {
-    var data = ''
-    res.on('data', function (chunk) {
-      data += chunk
-    })
-    res.on('end', function () {
-      event.sender.send('has_filter_list', data)
-    })
-  }).on('error', function (e) {
-    event.sender.send('has_filter_list', '{"error":true, "code":"could_not_connect", "message":"Could not connect to server"}')
-  }).end()
-})
-
-ipc.on('filter_delete', function (event, arg) {
-  var options = {
-    hostname: 'localhost',
-    port: settings.getPort() || 3000,
-    path: '/mapfilter/filters/local/' + arg,
-    method: 'DELETE'
-  }
-  http.request(options, function (res) {
-    var data = ''
-    res.on('data', function (chunk) {
-      data += chunk
-    })
-    res.on('end', function() {
-      event.sender.send('filter_list_changed', data);
-    })
-  }).on('error', function (e) {
-    event.sender.send('filter_list_changed');
-  }).end()
-})
-
-ipc.on('settings_list', function (event, arg) {
-  settings.get(function(err, selected) {
-    var defaults = settings.getDefaults();
-    var workspaces = settings.getWorkspaces();
-    if (!err && selected) {
-      for (var key in selected)
-        defaults[key] = selected[key]
-    }
-    workspaces.current = settings.getWorkspace();
-    event.sender.send('has_settings_list', {settings: defaults, workspaces: workspaces});
-  })
-})
-
-ipc.on('settings_save', function (event, arg) {
-  settings.patch(arg, function (err) {
-    if (err) {
-      event.sender.send('has_settings_save', '{"error",true, "code":"could_not_save_settings", "message":"Could not save settings file"}')
-    } else {
-      event.sender.send('has_settings_save', '{"error":false}')
-    }
-  })
-})
-
-ipc.on('workspace_create', function (event, arg) {
-  workspaces.create(arg, function(err, result) {
-    if (err)
-      event.sender.send('on_workspace_changed', err);
-    else
-      event.sender.send('on_workspace_changed', { error: false, result_message: 'workspace_created', workspace: result });
-  });
-});
-
-ipc.on('workspace_change', function (event, arg) {
-  workspaces.setCurrent(arg, function(err, result) {
-    if (err)
-      event.sender.send('on_workspace_changed', err);
-    else
-      event.sender.send('on_workspace_changed', { error: false, result_message: 'workspace_changed' });
-  });
-});
-
-ipc.on('workspace_delete', function (event, arg) {
-  workspaces.remove(arg, function(err, result) {
-    if (err)
-      event.sender.send('on_workspace_changed', err);
-    else
-      event.sender.send('on_workspace_changed', { error: false, result_message: 'workspace_deleted', workspace: arg });
-  });
-});
-
-ipc.on('templates_list', function (event, arg) {
-  var folder = path.join(__dirname, 'templates');
-  glob(path.join(folder, "**/template.json"), function(err, files) {
-    var list = [];
-    if (!err) {
-      for (var i in files) {
-        try {
-          var json = JSON.parse(fs.readFileSync(files[i], 'utf8'));
-          json["id"] = path.basename(path.dirname(files[i]));
-          list.push(json);
-        } catch (e) {
-          //continue;
-        }
-      }
-    }
-    event.sender.send('has_templates_list', list);
-  });
-});
-
-/*ipc.on('save_template', function (event, arg) {
-  var content_dir = path.join(path.dirname(settings.getSubmissionsDirectory()), 'content')
-  fs.mkdirpSync(content_dir)
-  site_builder.build_website({
-    content: content_dir,
-    target: 'website',
-    template: path.join('templates', arg.template),
-    context: {
-      'editable': true,
-      'environment': 'monitoring-station'
-    },
-    parameters: arg,
-    callback: function () {
-      event.sender.send('has_saved_template', '{"error":false}')
-    }
-  })
-})*/
-
-ipc.on('list_map_layers', function (event, arg) {
-  listTiles(function(err, result) {
-    if (err)
-      event.sender.send('has_list_map_layers', [])
-    else {
-      event.sender.send('has_list_map_layers', result.tiles.filter(function(value) {
-        return value.valid;
-      }).map(function(value) {
-        return {name: value.name, value: value.key + ''}
-      }));
-    }
-  });
-});
-
-ipc.on('import_files', function(event, args) {
-  var options = {
-    properties: ['openFile'],
-    filters: [ { name: 'ZIP', extensions: ['zip'] } ]
-  };
-
-  dialog.showOpenDialog(mainWindow, options, function(file) {
-    if (file) {
-      var source = '' + file;
-      var target = path.dirname(settings.getSubmissionsDirectory());
-
-      var complete = function(err) {
-        if (err) {
-          event.sender.send('has_import_files', { error: true, code: 'import_delete_failed', ex: err });
-        } else {
-          try {
-            fs.createReadStream(source)
-            .pipe(unzip.Extract({ path: target }))
-            .on('close', function() {
-              GeoJson.generate(function(err_json, details) {
-                if (err_json)
-                  event.sender.send('has_import_files', err_json);
-                else
-                  event.sender.send('has_import_files', { error: false, details: details });
-              });
-            });
-          } catch (e) {
-            event.sender.send('has_import_files', { error: true, code: 'import_unzip_failed', ex: e });
-          }
-        }
-      };
-
-      if (args.mode == 'merge')
-        complete(null);
-      else
-        fs.emptydir(path.join(target, 'Submissions'), complete);
-    } else {
-      event.sender.send('has_import_files', { error: true, code: 'import_cancelled' });
-    }
-  });
-});
-
-ipc.on('community_lands_backup', function(event, arg) {
-  if (settings.getCommunityLandsServer()) {
-    var options = {
-      hostname: 'localhost',
-      port: settings.getPort() || 3000,
-      path: '/communitylands/latest?submissions=true&website=true',
-      method: 'GET'
-    }
-    http.request(options, function (res) {
-      var data = ''
-      res.on('data', function (chunk) {
-        data += chunk
-      })
-      res.on('end', function () {
-        event.sender.send('has_community_lands_backup', data)
-      })
-    }).on('error', function (e) {
-      event.sender.send('has_community_lands_backup', '{"error":true, "code":"could_not_connect", "message":"Could not connect to server"}')
-    }).end()
-  } else {
-    event.sender.send('has_community_lands_backup', '{"error":true, "code":"community_lands_not_configured", "message":"Community Lands connection not configured"}')
-  }
-})
-
-ipc.on('community_lands_status', function (event, arg) {
-  if (settings.getCommunityLandsServer()) {
-    var options = {
-      hostname: 'localhost',
-      port: settings.getPort() || 3000,
-      path: '/communitylands/status',
-      method: 'GET'
-    }
-    http.request(options, function (res) {
-      var data = ''
-      res.on('data', function (chunk) {
-        data += chunk
-      })
-      res.on('end', function () {
-        event.sender.send('has_community_lands_status', data)
-      })
-    }).on('error', function (e) {
-      event.sender.send('has_community_lands_status', null)
-    }).end()
-  } else {
-    event.sender.send('has_community_lands_status', null)
-  }
-})
-
-ipc.on('community_lands_online', function (event, arg) {
-  event.sender.send('has_community_lands_online', arg)
-})
-
-var FiltersDir = settings.getFiltersDirectory()
-try {
-  fs.mkdirpSync(FiltersDir)
-} catch (e) { // It's ok
-}
-try {
-  fs.watch(FiltersDir, function (evt, filename) {
-    mainWindow.send('filter_list_changed');
-  })
-} catch (e) { // It's ok
-}
-
-function setupTilesWatcher() {
-  if (TilesWatcher)
-    TilesWatcher.close();
-
-  TilesWatcher = new watcher.FSWatcher({
-    //ignored: /(^|[\/\\])\../,
-    ignoreInitial: true,
-    persistent: true,
-    depth: 1
-  });
-
-  var TilesDirs = settings.getTilesDirectories();
-  for (var i in TilesDirs) {
-    var TilesDir = TilesDirs[i];
-    if (TilesDir == settings.getDefaultTilesDirectory()) {
-      try {
-        fs.ensureDirSync(TilesDir)
-      } catch (e) { // It's ok
-      }
-    }
-    try {
-      TilesWatcher.add(TilesDir)
-    } catch (e) { // It's ok
-    }
-  }
-  TilesWatcher.on('raw', function(evt, path) {
-    if (TilesWatcherTimeout)
-      clearTimeout(TilesWatcherTimeout)
-    TilesWatcherTimeout = setTimeout(refreshTiles, 1000);
-  });
-}
-setupTilesWatcher();
-
-function refreshTiles() {
-  Tiles.refresh(function() {
-    mainWindow.send('tiles_list_changed');
-  });
-}
-
-var TracksDir = settings.getTracksDirectory();
-try {
-  fs.ensureDirSync(TracksDir);
-} catch (e) { // It's ok
-}
-try {
-  fs.watch(TracksDir, { recursive: true }, function() {
-    mainWindow.send('tracks_list_changed');
-  });
-} catch (e) { //It's ok
-}
-
-function isInt(value) {
-  var check = parseInt(value)
-  return (!isNaN(value) && (check | 0) === check)
-}
+/** Show the application */
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
